Create a "next" parameter for signing up, and tunnel that "next" parameter through the entire registration/activation process.
