+++ kalite/main/middleware.py
@@ -5,15 +5,6 @@ from config.models import Settings
 from settings import LOG as logging
 
 
-class GetNextParam:
-    def process_request(self, request):
-        next = request.GET.get("next", "")
-        if next.startswith("/"):    
-            request.next = next
-        else:
-            request.next = ""
-
-
 # TODO(dylan): new class that handles finding and setting the language for the session
 class SessionLanguage:
     def process_request(self, request):
