+++ kalite/middleware.py
@@ -1,0 +1,11 @@
+from settings import LOG as logging
+
+class GetNextParam:
+    def process_request(self, request):
+        next = request.GET.get("next", "")
+        if next.startswith("/"):
+            logging.debug("next='%s'" % next)
+            request.next = next
+        else:
+            request.next = ""
+
