+++ kalite/templates/registration/activation_email.txt
@@ -1,6 +1,6 @@
 {% load humanize %}
 Someone, hopefully you, signed up for a new account with KA Lite using this email address. If it was you, and you'd like to activate and use your account, click the link below or copy and paste it into your web browser's address bar:
 
-http://{{ central_server_host }}/accounts/activate/{{ activation_key }}/
+http://{{ central_server_host }}/accounts/activate/{{ activation_key }}/{% if next %}?next={{ next }}{% endif %}
 
 If you didn't request this, you don't need to do anything; you won't receive any more email from us, and the account will expire automatically in {{ expiration_days|apnumber }} days.
